Refactored the classes that implement entity to belong to an entities package
