+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/api/BackupImportController.java
@@ -6,10 +6,10 @@ import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomField;
 import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomFieldRepository;
 import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomFieldRequestDto;
 import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomFieldValue;
-import com.sethhaskellcondie.thegamepensiveapi.domain.system.SystemGateway;
-import com.sethhaskellcondie.thegamepensiveapi.domain.system.SystemRequestDto;
-import com.sethhaskellcondie.thegamepensiveapi.domain.toy.ToyGateway;
-import com.sethhaskellcondie.thegamepensiveapi.domain.toy.ToyRequestDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemGateway;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemRequestDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyGateway;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyRequestDto;
 import com.sethhaskellcondie.thegamepensiveapi.domain.exceptions.ExceptionBackupImport;
 import com.sethhaskellcondie.thegamepensiveapi.domain.exceptions.ExceptionResourceNotFound;
 import org.springframework.web.bind.annotation.PostMapping;
