+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/api/HeartbeatController.java
@@ -4,8 +4,8 @@ import com.sethhaskellcondie.thegamepensiveapi.domain.Keychain;
 import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomField;
 import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomFieldRepository;
 import com.sethhaskellcondie.thegamepensiveapi.domain.customfield.CustomFieldValue;
-import com.sethhaskellcondie.thegamepensiveapi.domain.system.SystemGateway;
-import com.sethhaskellcondie.thegamepensiveapi.domain.toy.ToyGateway;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemGateway;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyGateway;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
