+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/entity/EntityGateway.java
@@ -1,4 +1,4 @@
-package com.sethhaskellcondie.thegamepensiveapi.domain;
+package com.sethhaskellcondie.thegamepensiveapi.domain.entity;
 
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.FilterRequestDto;
