+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/entity/toy/ToyService.java
@@ -1,11 +1,11 @@
-package com.sethhaskellcondie.thegamepensiveapi.domain.toy;
+package com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy;
 
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityService;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityService;
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.FilterService;
 import org.springframework.stereotype.Service;
 
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityRepository;
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityServiceAbstract;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityRepository;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityServiceAbstract;
 
 @Service
 public class ToyService extends EntityServiceAbstract<Toy, ToyRequestDto, ToyResponseDto> implements EntityService<Toy, ToyRequestDto, ToyResponseDto> {
