+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/filter/GetWithFiltersBooleanAndTimeTests.java
@@ -1,8 +1,8 @@
 package com.sethhaskellcondie.thegamepensiveapi.domain.filter;
 
 import com.sethhaskellcondie.thegamepensiveapi.domain.Keychain;
-import com.sethhaskellcondie.thegamepensiveapi.domain.system.System;
-import com.sethhaskellcondie.thegamepensiveapi.domain.system.SystemRepository;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.System;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
