+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/toy/ToyEntityTests.java
@@ -1,5 +1,8 @@
 package com.sethhaskellcondie.thegamepensiveapi.domain.toy;
 
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.Toy;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyRequestDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyResponseDto;
 import com.sethhaskellcondie.thegamepensiveapi.domain.exceptions.ExceptionMalformedEntity;
 import org.junit.jupiter.api.Test;
