+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/toy/ToyGatewayTests.java
@@ -1,6 +1,11 @@
 package com.sethhaskellcondie.thegamepensiveapi.domain.toy;
 
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityGatewayTests;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityGatewayTests;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.Toy;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyGateway;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyRequestDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyResponseDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyService;
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.Filter;
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.FilterRequestDto;
