+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/toy/ToyServiceTests.java
@@ -1,7 +1,11 @@
 package com.sethhaskellcondie.thegamepensiveapi.domain.toy;
 
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityRepository;
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityServiceTests;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityRepository;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityServiceTests;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.Toy;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyRequestDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyResponseDto;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyService;
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.FilterService;
 
 import static org.mockito.Mockito.mock;
