+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/toy/ToyWebTestClientTests.java
@@ -3,6 +3,7 @@ package com.sethhaskellcondie.thegamepensiveapi.domain.toy;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sethhaskellcondie.thegamepensiveapi.domain.entity.toy.ToyResponseDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
